"Added properties for NuGet package generation"
Added several properties to the project file to enable NuGet package generation. The `<GeneratePackageOnBuild>` property has been set to `True` to generate a NuGet package every time the project is built. The `<PackageId>` property has been set to `Tranborg.$(AssemblyName)` to serve as the unique identifier for the generated package. Lastly, the `<Title>` property has been set to `Traborg` to be the title of the generated NuGet package.